+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Read config.ini for mods/data paths. Removed dataPath in launch config, added configPath. modsPath and configPath also set the corresponding command line args.
 * Don't recheck info.json during a debug session, only when starting a new session.
 * Fixed crash when running remote.call() in /c
+* Don't attempt to catch an exception if there is no locatable source to show it at (eg. console commands)
 
 ## 0.18.7